Support older SDKs w/o templates
