+++ pebble_tool/sdk/templates/appinfo.json
@@ -1,0 +1,18 @@
+{
+  "uuid": "${uuid}",
+  "shortName": "${project_name}",
+  "longName": "${project_name}",
+  "companyName": "MakeAwesomeHappen",
+  "versionLabel": "1.0",
+  "sdkVersion": "${sdk_version}",
+  "targetPlatforms": ["aplite", "basalt", "chalk"],
+  "watchapp": {
+    "watchface": false
+  },
+  "appKeys": {
+    "dummy": 0
+  },
+  "resources": {
+    "media": []
+  }
+}
