+++ pebble_tool/sdk/templates/gitignore
@@ -1,0 +1,5 @@
+# Ignore build generated files
+build
+
+# Ignore waf lock file
+.lock-waf*
