+++ pebble_tool/sdk/templates/pebble-js-app.js
@@ -1,0 +1,5 @@
+Pebble.addEventListener("ready",
+    function(e) {
+        console.log("Hello world! - Sent from your javascript application.");
+    }
+);
