+++ pebble_tool/sdk/templates/simple.c
@@ -1,0 +1,5 @@
+#include <pebble.h>
+
+int main(void) {
+  app_event_loop();
+}
