+++ pebble_tool/sdk/templates/worker.c
@@ -1,0 +1,5 @@
+#include <pebble_worker.h>
+
+int main(void) {
+  worker_event_loop();
+}
